Better Collission Boxes ! ! ! (Sorta)
Decided to make the collission box vector into a 2 dimensional array. The columns are the sides of the box, and the rows are the entity the box is part of. This also makes it easier to automate the process of making collission boxes, now I can just create a new entity and it will automatically assign it a collission box. Next time I'm gonna have to figure out stopping any entity— that isn't controlled by the player— from noclipping with objects when they collide with them. This is gonna be hard